modified:   src/ARF.java 	modified:   src/RRF.java 	modified:   src/ReservationStation.java

+++ Lab-3/src/ARF.java
@@ -6,7 +6,7 @@ public class ARF {
     public class Register {
         public boolean busy = false;
         public int tag = -1;
-        public double value;
+        public int value;
     }
 
     public Register[] registers = new Register[Global.NUM_REGISTERS];
